backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alpaca-trade-api==3.2.0
alpha_vantage==3.0.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cmdstanpy==1.2.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
currents==0.0.1
cycler==0.12.1
Cython==3.0.11
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.114.1
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
h5py==3.13.0
holidays==0.58
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.8
iexfinance==0.5.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.9.2
kiwisolver==1.4.7
kubernetes==30.1.0
libclang==18.1.1
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.3
multidict==6.3.2
multitasking==0.0.11
namex==0.0.8
narwhals==1.9.1
newsapi-python==0.2.7
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.44.1
openrouter==1.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pillow==10.4.0
pinecone==5.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
plotly==5.24.1
posthog==3.6.5
propcache==0.3.1
prophet==1.1.6
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.11
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
stanio==0.5.1
starlette==0.38.5
streamlit==1.39.0
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.50.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.30.6
vaderSentiment==3.3.2
watchdog==5.0.3
watchfiles==0.24.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.44
zipp==3.20.1
